// src/types/ProjectDetails.ts
export interface ProjectDetails {
  id: string;
  title: string;
  description: string;
  images: string[];
  technologies: string[];
  duration: string;
  team: string;
  role: string;
  liveDemoLink: string;
  frontendCodeLink: string;
  backendCodeLink: string;
  overview: string;
  keyFeatures: string[];
  challengesAndSolutions: string;
}